Add support for inlining entire sub-template documents having a single top-level para
